test: [rebase] Move details of operation test from page tests to form tests
