+++ bciers/apps/administration/tests/components/operations/OperationInformationForm.test.tsx
@@ -1,7 +1,15 @@
-import { act, fireEvent, render, screen } from "@testing-library/react";
+import { act, fireEvent, render, screen, within } from "@testing-library/react";
 import { RJSFSchema } from "@rjsf/utils";
 import OperationInformationForm from "apps/administration/app/components/operations/OperationInformationForm";
 import { actionHandler, useSession } from "@bciers/testConfig/mocks";
+import {
+  getBusinessStructures,
+  getNaicsCodes,
+  getOperationWithDocuments,
+  getRegulatedProducts,
+  getReportingActivities,
+} from "./mocks";
+import { createAdministrationOperationInformationSchema } from "apps/administration/app/data/jsonSchema/operationInformation/administrationOperationInformation";
 
 useSession.mockReturnValue({
   data: {
@@ -11,8 +19,51 @@ useSession.mockReturnValue({
   },
 });
 
+export const fetchFormEnums = () => {
+  // Naics codes
+  getNaicsCodes.mockResolvedValue([
+    {
+      id: 1,
+      naics_code: "211110",
+      naics_description: "Oil and gas extraction (except oil sands)",
+    },
+    {
+      id: 2,
+      naics_code: "212114",
+      naics_description: "Bituminous coal mining",
+    },
+  ]);
+  // Reporting activities
+  getReportingActivities.mockResolvedValue([
+    {
+      id: 1,
+      name: "General stationary combustion excluding line tracing",
+      applicable_to: "all",
+    },
+    {
+      id: 2,
+      name: "Fuel combustion by mobile equipment",
+      applicable_to: "sfo",
+    },
+  ]);
+
+  // Business structures
+  getBusinessStructures.mockResolvedValue([
+    { name: "General Partnership" },
+    { name: "BC Corporation" },
+  ]);
+
+  // Regulated products
+  getRegulatedProducts.mockResolvedValue([
+    { id: 1, name: "BC-specific refinery complexity throughput" },
+    { id: 2, name: "Cement equivalent" },
+  ]);
+
+  // Registration purposes
+  actionHandler.mockResolvedValue(["Potential Reporting Operation"]);
+};
+
 // Just using a simple schema for testing purposes
-// OperationInformationPage test will test the proper schema with extra enums provided by the backend
 const testSchema: RJSFSchema = {
   type: "object",
   properties: {
@@ -103,6 +154,28 @@ const testSchemaWithOpt: RJSFSchema = {
 const formData = {
   name: "Operation 3",
   type: "Single Facility Operation",
+  naics_code_id: 1,
+  secondary_naics_code_id: 2,
+  operation_has_multiple_operators: true,
+  registration_purposes: ["Non-Regulated"],
+  activities: [1, 2, 3, 4, 5],
+  multiple_operators_array: [
+    {
+      mo_is_extraprovincial_company: false,
+      mo_legal_name: "Videl",
+      mo_trade_name: "Saiyaman 2",
+      mo_business_structure: "BC Corporation",
+      mo_cra_business_number: "0123456789",
+      mo_bc_corporate_registry_number: "abc0123456",
+      mo_attorney_street_address: "test st",
+      mo_municipality: "Victoria",
+      mo_province: "BC",
+      mo_postal_code: "V1V1V1",
+    },
+  ],
+  registration_purpose: "Non-Regulated",
+  regulated_products: [6],
+  opt_in: false,
 };
 
 const optInFormData = {
@@ -126,6 +199,9 @@ const optInFormData = {
 const operationId = "8be4c7aa-6ab3-4aad-9206-0ef914fea063";
 
 describe("the OperationInformationForm component", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
   it("renders the OperationInformationForm component", async () => {
     render(
       <OperationInformationForm
@@ -143,16 +219,57 @@ describe("the OperationInformationForm component", () => {
   });
 
   it("should render the form with the correct values when formData is provided", async () => {
+    fetchFormEnums();
+    const createdFormSchema =
+      await createAdministrationOperationInformationSchema(
+        formData.registration_purposes,
+        false,
+      );
     render(
       <OperationInformationForm
         formData={formData}
-        schema={testSchema}
+        schema={createdFormSchema}
         operationId={operationId}
       />,
     );
+    //name
+    expect(screen.getByText(/Operation 3/i)).toBeVisible();
+    // type
+    expect(screen.getByText(/Single Facility Operation/i)).toBeVisible();
+    // primary naics code
+    expect(screen.getByText(/Oil and gas extraction/i)).toBeVisible();
+    // secondary
+    expect(screen.getByText(/Bituminous coal mining/i)).toBeVisible();
+    // reporting activities
+    expect(screen.getByText(/Reporting Activities/i)).toBeVisible();
+    expect(screen.getByText(/General stationary combustion/i)).toBeVisible();
+    // 3 file inputs
+    expect(screen.getByText(/Process Flow Diagram/i)).toBeVisible();
+    expect(screen.getByText(/Boundary Map/i)).toBeVisible();
+    expect(screen.getByText(/Equipment List/i)).toBeVisible();
+    expect(screen.getAllByText(/No attachment was uploaded/i)).toHaveLength(3);
+    // multiple operators
+    expect(
+      screen.getByText(/Does the operation have multiple operators/i),
+    ).toBeVisible();
+    expect(screen.getByText(/Yes/i)).toBeVisible();
+    expect(screen.getByText(/Videl/i)).toBeVisible(); // legal name
+    expect(screen.getByText(/Saiyaman 2/i)).toBeVisible(); // trade name
+    expect(screen.getByText(/BC Corporation/i)).toBeVisible(); // business structure
+    expect(screen.getByText(/0123456789/i)).toBeVisible(); // business number
+    expect(screen.getByText(/abc0123456/i)).toBeVisible(); // registry number
+    expect(screen.getByText(/test st/i)).toBeVisible(); // adress
+    expect(screen.getByText(/Victoria/i)).toBeVisible(); //municipality
+    expect(screen.getByText(/British Columbia/i)).toBeVisible(); // province
+    expect(screen.getByText(/V1V1V1/i)).toBeVisible(); // postal code
 
-    expect(screen.getByText(/Operation Name/i)).toBeVisible();
-    expect(screen.getByText(/Operation Type/i)).toBeVisible();
+    // registration info & purpose
+    expect(
+      screen.getByText(
+        /The purpose of this registration is to register as a\:/i,
+      ),
+    ).toBeVisible();
+    expect(screen.getByText(/Non-Regulated/i)).toBeVisible();
   });
 
   it("should enable editing when the Edit button is clicked", async () => {
@@ -284,7 +401,7 @@ describe("the OperationInformationForm component", () => {
       screen.getByRole("button", { name: "Submit" }).click();
     });
 
-    expect(actionHandler).toHaveBeenCalledTimes(3);
+    expect(actionHandler).toHaveBeenCalledTimes(2);
     expect(actionHandler).toHaveBeenCalledWith(
       `registration/v2/operations/${operationId}/registration/opted-in-operation-detail`,
       "PUT",
@@ -311,5 +428,125 @@ describe("the OperationInformationForm component", () => {
         expect(radioValue).toHaveTextContent(/no/i);
       });
     });
+  });
+
+  it("should render the opt-in information if purpose is opt-in", async () => {
+    getOperationWithDocuments.mockResolvedValueOnce(optInFormData);
+
+    render(
+      <OperationInformationForm
+        formData={optInFormData}
+        schema={testSchemaWithOpt}
+        operationId={operationId}
+      />,
+    );
+
+    expect(screen.getByText(/Operation 5/i)).toBeVisible();
+    // Questions - 1 (borrowed from operations form tests)
+    const meetsSection3EmissionsRequirementsView = screen.getByText(
+      /does this operation have emissions that are attributable for the purposes of section 3 of \?/i,
+    );
+    within(meetsSection3EmissionsRequirementsView).getByRole("link", {
+      name: /the act/i,
+    });
+    expect(
+      screen.getAllByRole("link", {
+        name: /the act/i,
+      })[0],
+    ).toHaveAttribute(
+      "href",
+      "https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/14029_01",
+    );
+
+    // Questions - 2
+    expect(
+      screen.getByText(/is this operation an electricity import operation\?/i),
+    ).toBeVisible();
+
+    // Questions - 3
+    expect(
+      screen.getByText(
+        /designation as an opt-in can only be granted to an entire operation \(i\.e\. not a part or certain segment of an operation\)\. do you confirm that the operation applying for this designation is an entire operation\?/i,
+      ),
+    ).toBeVisible();
+
+    // Questions - 4
+    const meetsSection6EmissionsRequirementsView = screen.getByText(
+      /does this operation have emissions that are attributable for the purposes of section 6 of \?/i,
+    );
+
+    within(meetsSection6EmissionsRequirementsView).getByRole("link", {
+      name: /the act/i,
+    });
+    expect(
+      screen.getAllByRole("link", {
+        name: /the act/i,
+      })[1],
+    ).toHaveAttribute(
+      "href",
+      "https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/14029_01",
+    );
+
+    // Questions - 5
+    const meetsNaicsCode1122562ClassificationRequirementsView =
+      screen.getByText(
+        /is this operation’s primary economic activity classified by the following/i,
+      );
+
+    within(meetsNaicsCode1122562ClassificationRequirementsView).getByRole(
+      "link",
+      {
+        name: /naics code - 11, 22, or 562\?/i,
+      },
+    );
+
+    expect(
+      screen.getByRole("link", {
+        name: /naics code - 11, 22, or 562\?/i,
+      }),
+    ).toHaveAttribute(
+      "href",
+      "https://www.statcan.gc.ca/en/subjects/standard/naics/2022/v1/index",
+    );
+
+    // Questions - 6
+    const meetsProducingGgerScheduleA1RegulatedProductView = screen.getByText(
+      /does this operation produce a regulated product listed in table 1 of schedule a\.1 of \?/i,
+    );
+    within(meetsProducingGgerScheduleA1RegulatedProductView).getByRole("link", {
+      name: /the ggerr/i,
+    });
+
+    expect(
+      screen.getByRole("link", {
+        name: /the ggerr/i,
+      }),
+    ).toHaveAttribute(
+      "href",
+      "https://www.bclaws.gov.bc.ca/civix/document/id/lc/statreg/249_2015",
+    );
+
+    // Questions - 7
+    const meetsReportingAndRegulatedObligationsView = screen.getByText(
+      /is this operation capable of fulfilling the obligations of a reporting operation and a regulated operation under and the regulations\?/i,
+    );
+    within(meetsReportingAndRegulatedObligationsView).getByRole("link", {
+      name: /the act/i,
+    });
+    expect(
+      screen.getAllByRole("link", {
+        name: /the act/i,
+      })[2],
+    ).toHaveAttribute(
+      "href",
+      "https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/14029_01",
+    );
+
+    // Questions - 8
+    expect(
+      screen.getByText(
+        /will the operator notify the director as soon as possible if this operation ceases to meet any of the criteria for the designation of the operation as a reporting operation and a regulated operation\?/i,
+      ),
+    );
   });
 });
